f17: Move data.py to Models/Data.py
