+++ settings.py
@@ -1,5 +1,5 @@
 import json
-from data import Data
+from Model.Data import Data
 from Model.Command import Command
